zero-copy: fix two directly connncted NeutronGraphs
The output of 1st NeutronGraph is used as the input of the 2nd.
In this case, the tensor is allocated into buffer1 by
neutron_buffer_create(...) and pushed by datamover as output.
However, this update is not in buffer2 of the 2nd NeutronGraph.

Fixed by below copy:
in Eval after run, copy result from butter1 to tensor data

In the long run, we need the global neutron buffer management to support
the sharing across the different NeutronGraphs.

Signed-off-by: Forrest Shi <xuelin.shi@nxp.com>